меняет рыбу на моки в блоке пользователя
