+++ src/mock/user-data.js
@@ -1,0 +1,5 @@
+import {getRandomNumber} from '../util.js';
+
+export const generateUserInfo = () => {
+  return getRandomNumber(1000);
+};
